renaming file entity.ts to likemark.ts
