+++ src/app/models/likemark.ts
